Meant to add -p only, wound up in the weeds, in a good way :>
* adds -p argument which formats ipv6 addresses for use in postfix $mynetworks settings (encapsulate address in []s)

* adds -v option which turns debug mode on (off by default)

* converts the straight script into one with a __main__ entrypoint, so it can be used as a submodule in other scripts.

* adds support for arbitrary numbers of domains to check in one invocation, e.g. `python spf_dump.py _spf.google.com _spf.facebook.com twitter.com` with or without -p or -v

BUG?: facebook as of this commit has a broken spf record that produces a blank line due to a " " placed in the middle of a key: i" "p4:66.220.157.0/25. I have opted to email postmaster@facebook.com and hope they fix it rather than add even more complexity to deal with this broken record.